rename: useAppendUtterances hook location in PostComment
